3/n chore(kurtosis): reorganize e2e tests into subdirectories (#3043)
