+++ testing/e2e/preconf/flow_test.go
@@ -18,13 +18,11 @@
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
 // TITLE.
 
-//go:build e2e_preconf
+//go:build e2e
 
-package e2e_test
+package preconf_test
 
 import (
-	"testing"
-
 	"github.com/berachain/beacon-kit/testing/e2e/config"
 	"github.com/berachain/beacon-kit/testing/e2e/suite"
 )
@@ -42,72 +40,55 @@ const (
 	blocksAfterFallback = 10
 )
 
-// PreconfE2ESuite is a test suite for preconfirmation functionality.
-type PreconfE2ESuite struct {
-	suite.KurtosisE2ESuite
-}
-
-// SetupSuite sets up the suite with preconf configuration.
-func (s *PreconfE2ESuite) SetupSuite() {
-	s.SetupSuiteWithOptions(suite.WithPreconfConfig())
-}
-
-// TestPreconfE2ESuite runs the preconf test suite.
-func TestPreconfE2ESuite(t *testing.T) {
-	suite.Run(t, new(PreconfE2ESuite))
-}
-
-// TestPreconfEndToEnd verifies the complete preconfirmation flow:
-// 1. Sequencer serves payloads to whitelisted validators
-// 2. Validators fetch payloads from sequencer
-// 3. When sequencer is removed, validators fall back to local building
-func (s *PreconfE2ESuite) TestPreconfEndToEnd() {
-	// Wait for network to produce enough blocks for preconf to be exercised
+// TestSequencerFlow verifies the preconf pathway using ordered subtests.
+// NOTE: This test removes the sequencer service in SequencerFallback,
+// so it must run after TestPreconfTransactions (testify runs alphabetically).
+func (s *PreconfE2ESuite) TestSequencerFlow() {
+	// Wait for enough blocks so preconf has been exercised.
 	err := s.WaitForFinalizedBlockNumber(blocksToWait)
 	s.Require().NoError(err, "Network should reach finalized blocks")
 
-	sequencer := sequencerCLService
-	fetchers := []string{config.ClientValidator0}
-
-	// Step 1: Verify sequencer is serving payloads
+	// Verify sequencer is serving payloads.
 	s.Run("SequencerServesPayloads", func() {
-		logs, err := s.GetServiceLogs(sequencer)
+		logs, err := s.GetServiceLogs(sequencerCLService)
 		s.Require().NoError(err, "Should get sequencer logs")
 
 		found := suite.ContainsLogMessage(logs, sequencerServingLog)
 		s.Require().True(found,
 			"Sequencer (%s) should serve payloads to validators. "+
 				"Expected log message containing: %q",
-			sequencer, sequencerServingLog)
+			sequencerCLService, sequencerServingLog)
 	})
 
-	// Step 2: Verify validators fetch from sequencer
-	for _, validator := range fetchers {
-		validator := validator // capture for closure
-		s.Run("ValidatorFetches/"+validator, func() {
-			logs, err := s.GetServiceLogs(validator)
-			s.Require().NoError(err, "Should get validator logs for %s", validator)
+	// Verify validators fetch from sequencer.
+	s.Run("ValidatorFetches", func() {
+		for _, validator := range []string{config.ClientValidator0} {
+			validator := validator // capture for closure
+			s.Run(validator, func() {
+				logs, err := s.GetServiceLogs(validator)
+				s.Require().NoError(err, "Should get validator logs for %s", validator)
 
-			found := suite.ContainsLogMessage(logs, validatorFetchingLog)
-			s.Require().True(found,
-				"Validator (%s) should fetch payloads from sequencer. "+
-					"Expected log message containing: %q",
-				validator, validatorFetchingLog)
-		})
-	}
+				found := suite.ContainsLogMessage(logs, validatorFetchingLog)
+				s.Require().True(found,
+					"Validator (%s) should fetch payloads from sequencer. "+
+						"Expected log message containing: %q",
+					validator, validatorFetchingLog)
+			})
+		}
+	})
 
-	// Step 3: Remove sequencer and verify fallback to local building
-	s.Run("FallbackAfterSequencerRemoved", func() {
-		// Get current block before removing sequencer
+	// Remove sequencer and verify validators fall back to local building.
+	s.Run("SequencerFallback", func() {
+		// Get current block before removing sequencer.
 		currentBlock, err := s.RPCClient().BlockNumber(s.Ctx())
 		s.Require().NoError(err, "Should get current block number")
 
-		// Remove sequencer - simulates crash/unavailability
-		s.T().Logf("Removing sequencer (%s)...", sequencer)
-		err = s.RemoveService(sequencer)
+		// Remove sequencer -- simulates crash/unavailability.
+		s.T().Logf("Removing sequencer (%s)...", sequencerCLService)
+		err = s.RemoveService(sequencerCLService)
 		s.Require().NoError(err, "Should remove sequencer service")
 
-		// Wait for more blocks - validators must build locally now
+		// Wait for more blocks -- validators must build locally now.
 		targetBlock := currentBlock + blocksAfterFallback
 		s.T().Logf("Waiting for %d more blocks (current: %d, target: %d)...",
 			blocksAfterFallback, currentBlock, targetBlock)
@@ -115,12 +96,12 @@ func (s *PreconfE2ESuite) TestPreconfEndToEnd() {
 		err = s.WaitForFinalizedBlockNumber(targetBlock)
 		s.Require().NoError(err, "Network should continue producing blocks after sequencer removed")
 
-		// Verify network continued
+		// Verify network continued.
 		finalBlock, err := s.RPCClient().BlockNumber(s.Ctx())
 		s.Require().NoError(err, "Should get final block number")
 		s.Require().GreaterOrEqual(finalBlock, targetBlock,
 			"Network should have produced blocks after sequencer removed")
 
-		s.T().Logf("Network continued: block %d → %d (fallback working)", currentBlock, finalBlock)
+		s.T().Logf("Network continued: block %d -> %d (fallback working)", currentBlock, finalBlock)
 	})
 }
